CUDA: Convert fixed integer division to multiplication
Summary:
Using code from Hacker's Delight and tricks in https://gmplib.org/~tege/divcnst-pldi94.pdf , one can convert signed or unsigned integer division to a multiplication-and-shift operation, if the divisor is known in advance.

This is the case for things like our copy kernel, where the divisors are known at launch time.

Signed integer div/mod this way is a factor of 2.18x faster than the native compiler. For unsigned, it varies from -5% slower -> 5% faster. In inspecting the SASS, the compiler seems to be doing similar tricks, but I think the unsigned performance still lags because the compiler doesn't know that the `if (add_)` branch is the same taken for all threads. I will try massaging the code to get better results, or perhaps switch completely to PTX.

This is for code calculating both the quotient `n / d` and remainder `n % d` together. Calculating unsigned quotients via magic numbers is slightly faster than the native compiler technique alone.

```
380.74ms  168.77us           (1200 1 1)       (128 1 1)        16        0B        0B         -           -   Tesla K40m (0)         1         7  divideMagicSigned(int, int*, facebook::cuda::SignedDivisor) [357]
381.06ms  180.35us           (1200 1 1)       (128 1 1)        20        0B        0B         -           -   Tesla K40m (0)         1         7  divideMagicUnsigned(int, unsigned int*, facebook::cuda::UnsignedDivisor) [363]
381.40ms  368.79us           (1200 1 1)       (128 1 1)        34        0B        0B         -           -   Tesla K40m (0)         1         7  divideSigned(int, int*, int) [369]
381.92ms  167.61us           (1200 1 1)       (128 1 1)        25        0B        0B         -           -   Tesla K40m (0)         1         7  divideUnsigned(int, unsigned int*, unsigned int) [375]
```

My intent is to roll this into the copy kernel, which will also be used for the math kernel. This is a stand-alone utility class for this.

Test Plan: Added randomized cpu/gpu tests to make sure that this calculates a division.

Reviewed By: tulloch@fb.com

Subscribers: fjargsto, ntv, jhj

FB internal diff: D1819506

Signature: t1:1819506:1423349325:d18e6b9e63d660c4069d305dcd4ceb3015e3208f