Deterministic LookupTableGPU; other deterministic fp stuff
Summary:
This removes usage of atomicAdd() with LookupTableGPU and makes the code deterministic.

This is done by a warp-wide collision detection algorithm using shuffles. This allows a warp to determine if gradient additions are going to collide, and if so, will serialize the summation.

Removed all the "feature in dim 2" stuff from the CUDA kernels, since you don't need to repeat the code. A transposition on the input will achieve this instead. DRY.

Adds two collision detection routines as warp reductions in fbcode/cuda with a test. The collision mask version was originally used, but was slower than just serializing all execution if there was a collision, so I didn't choose to use it here.

Also adds an implementation of "Algorithm 5" from http://www.eecs.berkeley.edu/~hdnguyen/public/papers/ARITH21_Fast_Sum.pdf for deterministic floating point reduction. I decided not to go with this in the lookup table, but it is a useful strategy to have available for making future algorithms deterministic.

Refactored the warp reduction stuff to be functor-based, so there's less code repetition.

Peng Wang from NVIDIA has a SPMV-like implementation of the same algorithm that he just sent me, but that requires CUB which we don't have available, and has comparable perf for the small sizes we're interested in. I think that strategy is worth looking into for other versions.

Test Plan:
Added a strict determinism test for LookupTableGPU, and tests for the warp-wide collision reduction routines.

This doesn't appear to fix RNN testEndToEnd, but that's because it is using a precompiled model.log? @spchopra, you'll have to regenerate that once this is in?

```
[jhj@devgpu025 ~/fbcode/cuda] ~/fbcode/_build/dbg/deeplearning/projects/rnn-orthogonal/test/rnn_test.llar
>>>>>>>>> <TestFunction>
>>> <TestFunction>:testEndToEnd
[[ Loading dataset and dictionary ]]
loading a new random training shard
[================================ 29000/29048 ==============================>.] ETA: 47ms | Step: 0ms
Entropy: 8.617463  ::  Base Entropy: 8.600580...e/deeplearning/projects/rnn-orthogonal/test/rnn_test.lua:111: assertion failed!
Failed
```

Reviewed By: spchopra@fb.com

Subscribers: vikas, fjargsto, ntv, jhj, spchopra, wojciech, soumith, tudorb, kma, tulloch

FB internal diff: D1871736

Tasks: 5944118

Signature: t1:1871736:1424978312:88552a223ebc73b36c1cf5a88bbff3ee4f898faa