cuda cleanup for release
Summary:I'm starting an effort to put out the latest stuff I worked on in oss state.
My goal is to make a bunch of rapid changes and to only land a big stack of small-ish diffs so the review process can be as painless as possible.
This first diff does 3 things:
- cleanups the constant names we use and uniformizes with Marat's constants
- introduces minor c++11 where it makes sense
- reimplements FFT forward 8, 16 and 32 using a mix of the fbfft paper and the nvidia strategy. The final result (at the end of the stack) is a code that is within 10% of nvidia's, that is readable, generic wrt implicit padding and uses no shared memory at all.

A consequence is that I also went ahead and modified the memory layout for 8, 16, 32.
I'll cleanup and uniformize in a subsequent diff, just trying to slice things up for review.
I temporarily deactivated iterated FFT based convolutions until I get all my ducks in a row.

This diff gets us about 20% speedup over trunk

Differential Revision: D2290166

fb-gh-sync-id: 67c4b02f307f7f3741aa86279d2182475e2db624
shipit-source-id: 67c4b02f307f7f3741aa86279d2182475e2db624