cuda cleanup for release: Hermitian symmetry
Summary:Adds Hermitian symmetry by computing a 2N real FFT as a N complex FFT.
This is the first part of the job which uses N / 2 + 1 unpacked registers per thread.
The next diff adds another Hermitian symmetry scheme and makes it N / 2.
This also cleans up a bit the Twiddles.

Differential Revision: D2290193

fb-gh-sync-id: e1470f542cd5acec2c7a7afa8e52e0ba83267675
shipit-source-id: e1470f542cd5acec2c7a7afa8e52e0ba83267675