New FBMMSmall with less reuse but very good latency -> batch size = 1 + fallback cases
Summary:This significantly improves the performance of low latency convolutions.
Basically this trades off reuse for latency: no reuse in shared memory but load with ldg of float4.
This is driven by getting good performance in COCO for batch size of 1 as well as more generally good perf for low latency convolutions.
I don't have perf numbers to share atm but running the _stressTest cases in test_FFT.lua with nvprof will give a good overview.

For the proxy version of COCO, the target is roughly 1x512x768 convolutions of 72x72 by 14x14
I fix the 512 dimension and look at the effect of changing 768:

  size             cudnn      this diff
  512 x 64        39           15.5
  512 x 128      45           33
  512 x 256      77           96
  512 x 512      147         212
  512 x 768      227         323

The projected performance if I scaled linearly would be around 160 after dropping the recomputation of fft(weights).
It feels like I am hitting a working set size limit with those large convolutions.
The solution is to implement tiling on batch, input, output which will also be useful in general.
As usual with COCO, the easy way out is increasing batch sizes.

Please comment while I'm away, I hope to land this and open source all the stack as soon as I get back.

General tiling willl be done later.

Reviewed By: soumith

Differential Revision: D2325470

fb-gh-sync-id: 71e35940aeb08d14a5c4e096bd1ec496d9913eac
shipit-source-id: 71e35940aeb08d14a5c4e096bd1ec496d9913eac