FBMM improvements
Summary:Last batch of improvements using some experience from the low latency MM and applying to the regular MM.
Perf looks gooood, need to see what is up on Maxwell but basically it seems we can also drop the CGemm pieces.
This will be done in big red diffs later once we confirm everything looks fine on Maxwell.

The gist of this diff is I now get to 70% of compute efficiency and 65% are floating point operations.
It used to be around 45-50%

Reviewed By: soumith

Differential Revision: D2352539

fb-gh-sync-id: 993a370dcf22ac1d52a68100ae0a5817a46157c7
shipit-source-id: 993a370dcf22ac1d52a68100ae0a5817a46157c7